Fix broken DeploymentRepositoryService / DeploymentRepository encapsulation
