+++ ejb3/src/main/java/org/jboss/as/ejb3/iiop/RemoteObjectSubstitutionService.java
@@ -22,9 +22,9 @@
 
 package org.jboss.as.ejb3.iiop;
 
+import org.jboss.as.ejb3.deployment.DeploymentRepository;
 import org.jboss.as.ejb3.logging.EjbLogger;
 import org.jboss.as.ejb3.deployment.DeploymentModuleIdentifier;
-import org.jboss.as.ejb3.deployment.DeploymentRepository;
 import org.jboss.as.ejb3.deployment.EjbDeploymentInformation;
 import org.jboss.as.ejb3.deployment.ModuleDeployment;
 import org.jboss.ejb.client.AbstractEJBMetaData;
